refactor: reuse shared AI agent modules
