+++ functions/src/https/index.js
@@ -10,4 +10,5 @@ export * from './studyUpdate.js'
 export * from './studyStorage.js'
 export * from './studyAnswers.js'
 export * from './heuristicAgents.js'
+export * from './agentModels.js'
 export * from './studyAIGenerate.js'
